refactor: rename historically named write_buffer_server
